requirements.txt
torch==2.1.0
torchvision==0.16.0
matplotlib==3.7.1
streamlit==1.29.0
Pillow==10.0.0
